Extract *common code of of the TX history endpoint
* the "common" code is not yet common, but we are adding an endpoint that will return txs in the same format as the history endpoint, only the main response wrapper will be different.
